+++ src/routes/report_routes.js
@@ -1,9 +1,17 @@
-const { NoContent, Created, BadRequest, Ok } = require("../model/common/model");
+import { NoContent, Created, BadRequest, Ok, Unauthorized } from "../model/common/model.js";
 
-const reportService = require("../services/report_service");
+import reportService from "../services/report_service.js";
+
+import { isUserAuthorizedOrAdmin, isUserAdmin } from "../plugins/middleware.js";
 
 const routeReport = (app) => {
   app.post("/api/report", async (req, res) => {
+    const authorized = isUserAuthorizedOrAdmin(req.cookie.login_token, req.body.user_reporting);
+
+    if (!authorized) {
+      return Unauthorized(res);
+    }
+
     const inserted = await reportService.addReport(req.body, req.cookie.login_token);
 
     if (!inserted) return BadRequest(res);
@@ -12,6 +20,12 @@ const routeReport = (app) => {
   });
 
   app.get("/api/report", async (req, res) => {
+    const authorized = isUserAdmin(req.cookie.login_token, req.query.id);
+
+    if (!authorized) {
+      return Unauthorized(res);
+    }
+
     const report = await reportService.getReports(req.query, req.cookie.login_token);
 
     if (report == null || report.length === 0) return NoContent(res);
@@ -20,6 +34,12 @@ const routeReport = (app) => {
   });
 
   app.delete("/api/report", async (req, res) => {
+    const authorized = isUserAdmin(req.cookie.login_token, req.query.id);
+
+    if (!authorized) {
+      return Unauthorized(res);
+    }
+
     const deleted = await reportService.deleteReport(req.query.id, req.cookie.login_token);
 
     if (!deleted) return NoContent(res);
@@ -28,4 +48,4 @@ const routeReport = (app) => {
   });
 };
 
-module.exports = routeReport;
+export default routeReport;
